% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{allele_validator}
\alias{allele_validator}
\title{allele information validator}
\format{
An object of class \code{validator} of length 3.
}
\usage{
allele_validator
}
\description{
allele information validator
}
\keyword{internal}
