% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{assert_chr}
\alias{assert_chr}
\title{Check that input contains a CHR column with values that look like (human) contigs}
\usage{
assert_chr(data, on_error = on_error_options)
}
\arguments{
\item{data}{a dataframe that has a column CHR which will be validated}

\item{on_error}{if data does _not_ validate, whether to show the reasons and whether to throw:
all: show all the problems and throw an exception
none: don't show anything, but throw an exception
summary: show a summary and throw an exception
tell: just return TRUE if OK and FALSE if invalid}
}
\value{
TRUE if data is valid and FALSE if invalid (only when "on_error" == "tell")
}
\description{
Check that input contains a CHR column with values that look like (human) contigs
}
\examples{

assert_chr(data.frame(CHR=c("1","2"))) # TRUE

assert_chr(data.frame(contig="1"),on_error="tell") # FALSE: column name needs to be "CHR"

assert_chr(data.frame(CHR=c("chr1","chr2"))) # TRUE

# contigs should either all have "chr" or all not have it
assert_chr(data.frame(CHR=c("1", "chr2")), on_error="tell") # FALSE 

# value needs to be 1-22, X,Y or with "chr" prefix.
assert_chr(data.frame(CHR="hello"),on_error="tell") # FALSE  



}
