% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{assert_gwas}
\alias{assert_gwas}
\title{Validate a dataframe as a gwas}
\usage{
assert_gwas(data, on_error = on_error_options)
}
\arguments{
\item{data}{input data, a dataframe}

\item{on_error}{if data does _not_ validate, whether to show the reasons and whether to throw:
all: show all the problems and throw an exception
none: don't show anything, but throw an exception
summary: show a summary and throw an exception
tell: just return TRUE if OK and FALSE if invalid}
}
\value{
TRUE if data is valid and FALSE if invalid (only when "on_error" == "tell")
}
\description{
Validate that a dataframe contains values that are consistent with being a gwas.
}
\examples{

 assert_gwas(demo_data) # TRUE

 broken_data <- demo_data # make copy
 broken_data$POS[1] <- 0 # Zero is not a valid value for POS
 assert_gwas(broken_data, on_error="tell") # FALSE

}
