% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{assert_rsids}
\alias{assert_rsids}
\title{Check that input contains an rsid column with values that look like rsids}
\usage{
assert_rsids(data, on_error = on_error_options)
}
\arguments{
\item{data}{a dataframe that has a column rsid which will be validated}

\item{on_error}{if data does _not_ validate, whether to show the reasons and whether to throw:
all: show all the problems and throw an exception
none: don't show anything, but throw an exception
summary: show a summary and throw an exception
tell: just return TRUE if OK and FALSE if invalid}
}
\value{
TRUE if data is valid and FALSE if invalid (only when "on_error" == "tell")
}
\description{
Check that input contains an rsid column with values that look like rsids
}
\examples{
assert_rsids(data.frame(rsid=c("rs001101","rs00042"))) # TRUE
assert_rsids(data.frame(rsid=c("rs001101"))) # TRUE
assert_rsids(data.frame(rsid="rs001101")) # TRUE

assert_rsids(data.frame(rsid=c("001101","rs00042")), on_error="tell") # FALSE 



}
