% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{assert_valid_data}
\alias{assert_valid_data}
\title{Check that input is a dataframe that validates according to the validator and optionally
show which data doesn't validate}
\usage{
assert_valid_data(data, validator, on_error = on_error_options)
}
\arguments{
\item{data}{a dataframe which will be validated against the validator}

\item{validator}{a validator against which to validate the data}

\item{on_error}{if data does _not_ validate, whether to show the reasons and whether to throw:
all: show all the problems and throw an exception
none: don't show anything, but throw an exception
summary: show a summary and throw an exception
tell: just return TRUE if OK and FALSE if invalid}
}
\value{
TRUE if data is valid and FALSE if invalid (only when "on_error" == "tell")
}
\description{
Check that input is a dataframe that validates according to the validator and optionally
show which data doesn't validate
}
