% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{chr_validator}
\alias{chr_validator}
\title{contig validator}
\format{
An object of class \code{validator} of length 2.
}
\usage{
chr_validator
}
\description{
contig validator
}
\keyword{internal}
