% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Snippet of a gwas}
\format{
A data frame with 6 rows and 10 variables:
\describe{
  \item{rsid}{The rsid of the SNP} 
  \item{CHR}{The name of the reference sequence the SNP is on}  
  \item{EA}{The effect allele for the beta}   
  \item{NEA}{The non-effect allele for the beta}  
  \item{EAF}{The frequency of the effect allele as observed in the study }  
  \item{beta}{The association coefficient between the effect allele and the outcome measured } 
  \item{SE}{The size of the estimated standard error of beta}   
  \item{P}{The P-value for beta being different from zero}    
  \item{POS}{The (1-based) position of the SNP within its reference sequence, CHR } 
  ...
}
}
\source{
\url{???}
}
\usage{
demo_data
}
\description{
A small part of a gwas provided for testing and examples
}
\keyword{datasets}
