% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_search.R
\name{fill_gwas_unknown_rsids}
\alias{fill_gwas_unknown_rsids}
\title{Method to obtain rsids for records in a gwas where they are missing}
\usage{
fill_gwas_unknown_rsids(gwas, assembly = valid_references, validate = TRUE)
}
\arguments{
\item{gwas}{a dataframe containing the gwas with CHR POS NEA and EA}

\item{assembly}{one of "hg18", "hg19" (default), or "hg38".}

\item{validate}{a boolean indicating whether to validate the resulting gwas 
(TRUE by default) use FALSE for debugging.}
}
\value{
original input gwas with available rsids replacing rsids where possible
}
\description{
The method looks at all the rows for which rsid is NA and gets their rsid by
calling get_unknown_rsids_from_locus and merging the resulting rsids into the gwas
using merge_rsids_into_gwas
}
\examples{

any(is.na(demo_data$rsid)) # TRUE
fixed_gwas = fill_gwas_unknown_rsids(demo_data)
any(is.na(fixed_gwas$rsid)) # FALSE
nrow(fixed_gwas) == nrow(demo_data) # TRUE



}
