% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_and_write_exposure_data}
\alias{filter_and_write_exposure_data}
\title{Method to filter and save exposure data}
\usage{
filter_and_write_exposure_data(
  data,
  location_prefix = ".",
  pvalue_threshold,
  rare_threshold
)
}
\arguments{
\item{data}{a dataframe with possibly snps that we want. Expected columns are P (p-value) and EAF (effect allele frequency)
but data is expected to validate according to \code{\link{assert_gwas}}}

\item{location_prefix}{a prefix pointing to where to save the filtered data}

\item{pvalue_threshold}{a p-value threshold for filtering the data (will look at the `P` column)}

\item{rare_threshold}{a threshold on the allele frequency to avoid rare variants (will look at the `EAF` column and 1-EAF)}
}
\value{
the filtered dataframe
}
\description{
Method to filter and save exposure data
}
