% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy_search.R
\name{fix_proxy_alleles}
\alias{fix_proxy_alleles}
\title{Since proxies are provided with a "Correlated_Alleles" column that links the old alleles to the new alleles.
This method converts the alleles accordingly}
\usage{
fix_proxy_alleles(proxies, reverse_direction = FALSE)
}
\arguments{
\item{proxies}{the dataframe containing (at least) the following columns EA, NEA (alleles), Correlated_Alleles}

\item{reverse_direction}{a flag indicating that the replacement should occur in the reverse direction}
}
\value{
the same dataframe with the alleles in EA and NEA replaced according to Correlated_Alleles column
}
\description{
Since proxies are provided with a "Correlated_Alleles" column that links the old alleles to the new alleles.
This method converts the alleles accordingly
}
\examples{

demo_data
fix_proxy_alleles(data.frame(demo_data,
                             Correlated_Alleles=c("G=G,A=T",
                                                  "C=G,G=C",
                                                  "A=G,G=T",
                                                  "C=T,T=A",
                                                  "C=C,A=T",
                                                  "A=T,G=C")))

}
