% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_2smr_results}
\alias{get_2smr_results}
\title{Run Two-Sample MR on pre-processed data.}
\usage{
get_2smr_results(preprocessed_snps)
}
\arguments{
\item{preprocessed_snps}{a dataframe containing a collection of snps that have been prepared for use with two-sample MR. 
the expected columns are: 
rsid -- the name of the SNP
beta.exp -- the exposure beta
SE.exp -- the exposure standard error
EAF.exp -- the exposure effect allele frequency
EA.exp -- the exposure effect allele
NEA.exp -- the exposure non-effect allele frequency
P.exp -- the exposure p-value
beta.out -- the outcome beta
SE.out -- the outcome standard error
EAF.out -- the outcome effect allele frequency
EA.out -- the outcome effect allele
NEA.out -- the outcome non-effect allele frequency
P.out -- the outcome p-value}
}
\value{
a list of many MR outcomes and plots
}
\description{
Run Two-Sample MR on pre-processed data.
}
