% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_search.R
\name{get_rsid}
\alias{get_rsid}
\title{Get rsid name from position}
\usage{
get_rsid(
  chrom,
  pos,
  ref,
  alt,
  assembly = valid_references,
  cache_file = NULL,
  update_cache = TRUE
)
}
\arguments{
\item{chrom}{The name of the contig for the SNP (the "chr" prefix will be ignored if present)}

\item{pos}{The (1-based) position of the SNP}

\item{ref}{The reference allele (only SNPs are supported)}

\item{alt}{The alternate allele (only SNPs are supported)}

\item{assembly}{Which reference genome to use ("hg18", "hg19", or "hg38")}

\item{cache_file}{Which file to use for caching results. if NULL will not use cache}

\item{update_cache}{A boolean indicating whether to update the cache results.}
}
\value{
an rsid identifier of the position provided
}
\description{
This method combines the cached and API version of get_rsid
}
\examples{

get_rsid("9", 125711603, "C", "A") # "rs10760259"
get_rsid("9", 136155000, "T", "C") # "rs635634"
get_rsid("9", 136155000, "C", "T") # "rs635634"
get_rsid("chr6", 32007459, "A",	"G") # "rs12525076"
get_rsid("chr6", 32007459, "G",	"A") # "rs12525076"
get_rsid("6", 32007459, "A",	"G") # "rs12525076"


}
\keyword{internal}
