% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_search.R
\name{get_rsid_from_position}
\alias{get_rsid_from_position}
\title{Get rsid name from position}
\usage{
get_rsid_from_position(chrom, pos, ref, alt, assembly = valid_references)
}
\arguments{
\item{chrom}{The name of the contig for the SNP}

\item{pos}{The (1-based) position of the SNP}

\item{ref}{The reference allele (only SNPs are supported)}

\item{alt}{The alternate allele (only SNPs are supported)}

\item{assembly}{Which reference genome to use ("hg18", "hg19", or "hg38")}
}
\value{
an rsid identifier of the position provided
}
\description{
This method accesses the NCI's API and queries a position (CHR & POS) and variant (REF & ALT)
to obtain a rsid. If it fails, it will attempt to swap REF and ALT before giving up.
}
\examples{

get_rsid_from_position("9", 125711603,	"C",	"A") # "rs10760259"
get_rsid_from_position("9", 136155000, "T", "C") # "rs635634"
get_rsid_from_position("9", 136155000, "C", "T") # "rs635634"



}
