% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_search.R
\name{get_unknown_rsids_from_locus}
\alias{get_unknown_rsids_from_locus}
\title{Method to obtain rsids for records in a gwas where they are missing}
\usage{
get_unknown_rsids_from_locus(
  gwas,
  assembly = valid_references,
  validate = TRUE
)
}
\arguments{
\item{gwas}{a dataframe containing the gwas with CHR POS NEA and EA}

\item{assembly}{one of "hg18", "hg19" (default), "hg38".}

\item{validate}{a boolean indicating whether to validate the resulting gwas 
(TRUE by default) use FALSE for debugging.}
}
\value{
subset of input gwas with available rsids replacing rsids where they were originally missing. Does
not include all the columns in the original gwas, only CHR, POS, NEA, EA, rsid
}
\description{
The method looks at all the rows for which rsid is NA and gets their rsid by
calling merge_rsids_into_gwas and merging the resulting rsids into the gwas
}
\examples{

any(is.na(demo_data$rsid)) # TRUE
partial_gwas = get_unknown_rsids_from_locus(demo_data)
any(is.na(partial_gwas$rsid)) # FALSE
nrow(partial_gwas) == nrow(demo_data) # false


}
\keyword{internal}
