% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{gwas_types_validator}
\alias{gwas_types_validator}
\title{gwas column type validator}
\format{
An object of class \code{validator} of length 7.
}
\usage{
gwas_types_validator
}
\description{
gwas column type validator
}
\keyword{internal}
