% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{gwas_validator}
\alias{gwas_validator}
\title{locus information validator}
\format{
An object of class \code{validator} of length 10.
}
\usage{
gwas_validator
}
\description{
locus information validator
}
\keyword{internal}
