% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{locus_validator}
\alias{locus_validator}
\title{locus information validator}
\format{
An object of class \code{validator} of length 3.
}
\usage{
locus_validator
}
\description{
locus information validator
}
\keyword{internal}
