% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_search.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not in operator}
\usage{
lhs \%notin\% rhs
}
\arguments{
\item{lhs}{vector or NULL: the values to be matched. Long vectors are supported.}

\item{rhs}{vector or NULL: the values to be matched. Long vectors are not supported.}
}
\value{
A logical vector, indicating if a match was NOT located for each element of x: thus the values are TRUE or FALSE and never NA.
}
\description{
Negates the result of %in%
}
\keyword{internal}
