% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{palindromic}
\alias{palindromic}
\title{List of genotypes that are palindromic}
\format{
An object of class \code{character} of length 4.
}
\usage{
palindromic
}
\description{
List of genotypes that are palindromic
}
\keyword{internal}
