% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{proxy_allele_validator}
\alias{proxy_allele_validator}
\title{proxy allele information validator}
\format{
An object of class \code{validator} of length 2.
}
\usage{
proxy_allele_validator
}
\description{
proxy allele information validator
}
\keyword{internal}
