% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{proxy_stats_validator}
\alias{proxy_stats_validator}
\title{stats information validator}
\format{
An object of class \code{validator} of length 1.
}
\usage{
proxy_stats_validator
}
\description{
stats information validator
}
\keyword{internal}
