% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy_search.R
\name{prune_snps}
\alias{prune_snps}
\title{Prunes a collection of SNPs provided in a dataframe.
The method is as follows:}
\usage{
prune_snps(data_with_rsids, ld_pairs, r2_threshold = 0.05)
}
\arguments{
\item{data_with_rsids}{an input dataframe that includes column rsid}

\item{ld_pairs}{the result from calling get_LD_pairs matrix on the collection of rsids (and their contigs) in data_with_rsids}

\item{r2_threshold}{the (r^2) threshold to use for considering two snps to be in LD}
}
\value{
a subset of the original dataframe, pruned according to some method....
}
\description{
Prunes a collection of SNPs provided in a dataframe.
The method is as follows:
}
