% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy_search.R
\name{remove_linked_snps}
\alias{remove_linked_snps}
\title{Method to select snps so that they are not in LD.}
\usage{
remove_linked_snps(
  input_snps,
  pop,
  token = Sys.getenv("LDLINK_TOKEN"),
  prune_r2_threshold = 0.05
)
}
\arguments{
\item{input_snps}{a dataframe}

\item{pop}{the population codes (or vector thereof) in which the LD is calculated e.g. c("CEU"), c("CEU","YRI")}

\item{token}{is an access token to the nci API.
if you don't have a token go here: https://ldlink.nci.nih.gov/?tab=apiaccess}

\item{prune_r2_threshold}{a threshold (in LD r^2) for two SNPs to be considered "close"}
}
\value{
a list with the pruned dataframe, the list of removed SNPs, and three plots (if ggplot2 is available)
}
\description{
Method to select snps so that they are not in LD.
}
