% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy_search.R
\name{replace_alleles}
\alias{replace_alleles}
\title{Replaces the alleles in a GWAS according to the replacement string provided.}
\usage{
replace_alleles(alleles, replacement_string, reverse_direction = FALSE)
}
\arguments{
\item{alleles}{a list of alleles that will be transformed according to the "rule" 
in replacement_string}

\item{replacement_string}{a set of replacement rules in the format "<CHAR_OLD>=<CHAR_NEW>,..."
where the appearances of CHAR_OLD will be replaced by CHAR_NEW.}

\item{reverse_direction}{a flag indicating that the replacement should occur in the reverse direction}
}
\description{
Replaces the alleles in a GWAS according to the replacement string provided.
}
\examples{
replace_alleles(list("C", "G"), "C=A,G=T") # c("A", "T")
replace_alleles(list("C", "G"), "C=C,G=T") # c("C", "T")
replace_alleles(list("C", "G"), "C=A,G=G") # c("A", "G")
replace_alleles(list("C", "G"), "C=T,G=A") # c("T", "A")
replace_alleles(list("C", "T"), "C=A,T=T") # c("A", "T")
replace_alleles(list("C", "T"), "C=A,T=C") # c("A", "C")
replace_alleles(list("C", "T"), "C=T,T=C") # c("T", "C")

}
\keyword{internal}
