% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{required_headers}
\alias{required_headers}
\title{List of column headers that are required to be present in a GWAS (for this package)}
\format{
An object of class \code{character} of length 9.
}
\usage{
required_headers
}
\description{
List of column headers that are required to be present in a GWAS (for this package)
}
\keyword{internal}
