% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{rsid_validator}
\alias{rsid_validator}
\title{rsid information validator}
\format{
An object of class \code{validator} of length 2.
}
\usage{
rsid_validator
}
\description{
rsid information validator
}
\keyword{internal}
