% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{stats_validator}
\alias{stats_validator}
\title{stats information validator}
\format{
An object of class \code{validator} of length 2.
}
\usage{
stats_validator
}
\description{
stats information validator
}
\keyword{internal}
