% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{valid_alleles}
\alias{valid_alleles}
\title{List of strings that are valid alleles}
\format{
An object of class \code{character} of length 4.
}
\usage{
valid_alleles
}
\description{
List of strings that are valid alleles
}
\keyword{internal}
