% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{valid_contigs}
\alias{valid_contigs}
\title{List of valid names for a reference sequence/contig in this package}
\format{
An object of class \code{character} of length 24.
}
\usage{
valid_contigs
}
\description{
This only allows for contigs without the "chr" prefix. for the list containing chr see
\code{\link{valid_contigs_with_chr}}
}
\keyword{internal}
