% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{valid_contigs_with_chr}
\alias{valid_contigs_with_chr}
\title{List of valid names for a reference sequence/contig in this package}
\format{
An object of class \code{character} of length 24.
}
\usage{
valid_contigs_with_chr
}
\description{
This only allows for contigs with the "chr" prefix. For the list containing see
\code{\link{valid_contigs}}
}
\keyword{internal}
