% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{valid_references}
\alias{valid_references}
\title{List of human reference builds that can be used to find rsids}
\format{
An object of class \code{character} of length 3.
}
\usage{
valid_references
}
\description{
the order matters...the first one is the default value for methods that 
need assembly as input.
}
\keyword{internal}
