% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vin.R
\name{\%vin\%}
\alias{\%vin\%}
\title{vin operator}
\usage{
x \%vin\% table
}
\arguments{
\item{vector}{or NULL: the values to be matched}

\item{table}{vector or NULL: the values to be matched against.}
}
\value{
A vector of the same length as x. TRUE if the element in x is 
definitely in table, FALSE if not and NA is unknown (e.g. if table contains NA)
}
\description{
See \code{validate::\link[validate:vin]{\%vin\%}} for details.
}
\keyword{internal}
